backend/flask_app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
flask-cors
python-dateutil